mysql버전 추가 업데이트2

+++ spring01/src/main/java/com/example/spring01/service/BoardServiceImpl.java
@@ -35,6 +35,7 @@ public class BoardServiceImpl implements BoardService{
 	public Map<String, Object> boardList(Map<String, Object> map) throws Exception {
 		return boardDao.boardList(map); 
 	}
+	
 	@Override 
 	public void insertBoard(Map<String, Object> map, HttpServletRequest request) throws Exception { 
 		boardDao.insertBoard(map);
@@ -84,8 +85,8 @@ public class BoardServiceImpl implements BoardService{
         // 세션에 저장된 조회시간 검색
         // 최초로 조회할 경우 세션에 저장된 값이 없기 때문에 if문은 실행X
         if(session.getAttribute("update_time_"+map.get("IDX")) != null){
-                                // 세션에서 읽어오기
-            update_time = (Integer)session.getAttribute("update_time_" + map.get("IDX"));
+            // 세션에서 읽어오기
+            update_time = (Long) session.getAttribute("update_time_" + map.get("IDX"));
         }else {
         	// 시스템의 현재시간을 current_time에 저장
         	long current_time = System.currentTimeMillis();
